Movement Reworked
Movement has been mostly redone and collision is now handled through
code instead of GM's default.  The result is that the character no
longer gets "stuck" on walls and that pressing opposite directions at
the same time results in the idle sprite instead of a walking animation
not going anywhere.